server-ts: Convert routes/api/similar_notes
