+++ src/routes/routes.js
@@ -51,7 +51,7 @@ const bulkActionRoute = require('./api/bulk_action');
 const specialNotesRoute = require('./api/special_notes');
 const noteMapRoute = require('./api/note_map');
 const clipperRoute = require('./api/clipper');
-const similarNotesRoute = require('./api/similar_notes.js');
+const similarNotesRoute = require('./api/similar_notes');
 const keysRoute = require('./api/keys');
 const backendLogRoute = require('./api/backend_log');
 const statsRoute = require('./api/stats.js');
